Restructured code to make it more modular, cleaned it up a bit and made it PEP8 compliant. Window size is now based off of monitor resolution instead of an arbitrary constant. Also fixed typos.
