+++ README.md
@@ -1,4 +1,4 @@
 # pred-prey-boids
-This project is a boids implementatation in python complete with obstacles and a goal. The boids have goal seeking and collision avoidance in addition to basic behaviors.
+This project is a boids implementation in python complete with obstacles and a goal. The boids have goal seeking and collision avoidance in addition to basic behaviors.
 
 It is written in python 2.7 with pygame.
